+++ frontend/src/pages/DashboardPageV2/DashboardContainer/DashboardSettings/Variables/__tests__/applyVariableToPanelsPatch.test.ts
@@ -7,7 +7,7 @@ import {
 	buildApplyVariableToPanelsPatch,
 	buildSyncVariableToPanelsPatch,
 	getPanelIdsReferencingVariable,
-} from '../applyVariableToPanelsPatch';
+} from '../utils/applyVariableToPanelsPatch';
 
 /** Minimal builder-query panel wrapped in a CompositeQuery, with a given filter. */
 function compositePanel(filterExpression: string): unknown {
