+++ frontend/src/pages/DashboardPageV2/DashboardContainer/DashboardSettings/Variables/utils/variablePatchOps.ts
